pyproject.toml
[project]
name = "superset-mcp-bridge"
version = "0.1.0"
description = "MCP bridge connecting Claude Desktop to Apache Superset"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["superset", "mcp", "claude", "ai", "business-intelligence"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=0.9.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/superset-mcp-bridge"
Documentation = "https://github.com/YOUR_USERNAME/superset-mcp-bridge/blob/main/README.md"
Repository = "https://github.com/YOUR_USERNAME/superset-mcp-bridge"
Issues = "https://github.com/YOUR_USERNAME/superset-mcp-bridge/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
where = ["."]
include = ["main*"]